requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
clodsa==1.2.27
commentjson==0.9.0
cycler==0.11.0
Django==4.0.6
djangorestframework==3.13.1
fonttools==4.34.4
google==3.0.0
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-vision==3.1.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.0
imutils==0.5.4
joblib==1.1.0
keras==2.9.0
kiwisolver==1.4.4
lark-parser==0.7.8
mahotas==1.4.13
matplotlib==3.5.2
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.2.0
progressbar2==4.0.0
proto-plus==1.22.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.7.31
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.11
